feat: add Biome configuration as alternative linting toolchain
Add shareable Biome configuration that mirrors ESLint, Prettier, and
Stylelint rules as a faster, unified alternative. Biome is a Rust-based
tool that's 10-20x faster than Node.js tooling.

Changes:
- Add biome.json with comprehensive linting and formatting rules
- Add BIOME_COMPATIBILITY.md documenting unsupported features
- Update README.md to present two toolchain approaches:
  * ESLint + Prettier + Stylelint (modular, extensive plugins)
  * Biome (unified, fast alternative)
- Split test.run.ts into separate test files by tool
- Add test.run.biome.ts with Biome test suite
- Extract shared test utilities to test.helpers.ts
- Update package.json to include Biome as optional peer dependency
- Update tsconfig.build.json to exclude test files from compilation

The Biome config includes:
- JavaScript/TypeScript linting and formatting
- JSON support with strict parsing
- CSS support (standard CSS only, no SCSS)
- Experimental HTML support
- VCS integration with .gitignore support

Known limitations documented in BIOME_COMPATIBILITY.md:
- No SCSS support (standard CSS only)
- No Web Components or Lit plugin equivalents
- Limited HTML linting compared to @html-eslint
- No plugin system for custom rules

Test results:
- Biome tests: All passing (4/4 detection, 2/2 auto-fix)
- Prettier tests: All passing (3/3)
- Stylelint tests: All passing (2/2)
- ESLint tests: Node.js module resolution issue (pre-existing)
- Build: Successful
