+++ README.md
@@ -1,54 +1,71 @@
 # @enke.dev/lint
 
-## Install packages
+Opinionated linting and formatting configurations for TypeScript web projects. Choose between two approaches:
+
+1. **ESLint + Prettier + Stylelint** (modular, pluggable) - Traditional toolchain with extensive plugin ecosystem
+2. **Biome** (all-in-one, fast) - Unified Rust-based alternative that's 10-20x faster
+
+## Approach 1: ESLint + Prettier + Stylelint
+
+### Install packages
 
 Make sure to install the necessary peer dependencies:
 
 ```bash
-npm i -D @enke.dev/lint eslint prettier @awmottaz/prettier-plugin-void-html
+npm i -D @enke.dev/lint eslint prettier @awmottaz/prettier-plugin-void-html typescript jiti typescript-eslint
 ```
 
-For Typescript support, additionally install:
+### Prepare ESLint config
 
-```bash
-npm i -D jiti typescript typescript-eslint
-```
+Create an `eslint.config.ts` file in the root of your project:
 
-## Prepare Eslint config
-
-Create a `eslint.config.js` (or `eslint.config.ts`) file in the root of your project and add the following content:
-
-```js
+```ts
 import config from '@enke.dev/lint';
 
 export default config;
 ```
 
 > [!NOTE]
-> Using Typescript may requires the root directory where to find the `tsconfig.json` to be specified.\
-> Therefore, a convenience function `setTsConfigRootDir` is provided to configure this option globally.
+> Using TypeScript requires specifying the root directory where to find the `tsconfig.json`.\
+> A convenience function `setTsConfigRootDir` is provided to configure this option globally.
 
-### Extending a config
+#### Extending the config
 
-For example for setting up the Typescript parser, you can extend the base config like this:
+You can extend the base config to add custom rules or override defaults:
 
 ```ts
 import config, { setTsConfigRootDir } from '@enke.dev/lint';
 
-export default defineConfig([
+export default [
   // extend the base config
   ...config,
   // configure typescript parser to your needs
   setTsConfigRootDir(import.meta.dirname),
-]);
+  // add custom rules (e.g., restrict imports)
+  {
+    rules: {
+      'no-restricted-imports': [
+        'error',
+        {
+          patterns: [
+            {
+              group: ['@/internal/*'],
+              message: 'Import from @/public/* instead',
+            },
+          ],
+        },
+      ],
+    },
+  },
+];
 ```
 
 > [!TIP]
 > Extending configurations works the same way with all other configs provided by this package.
 
-### Using Monorepos
+#### Using Monorepos
 
-The VSCode Eslint plugin can be configured to pick up packages correctly by updating your `settings.json`, e.g.:
+The VSCode ESLint plugin can be configured to pick up packages correctly by updating your `settings.json`:
 
 ```json
 {
@@ -56,9 +73,9 @@ The VSCode Eslint plugin can be configured to pick up packages correctly by upda
 }
 ```
 
-## Prepare Prettier config
+### Prepare Prettier config
 
-A [shared prettier configuration](https://prettier.io/docs/sharing-configurations) can be used by creating a `prettier.config.js` (or `prettier.config.ts`) file in the root of your project with the following content:
+Create a `prettier.config.js` file in the root of your project:
 
 ```js
 import config from '@enke.dev/lint/prettier.config.js';
@@ -66,33 +83,118 @@ import config from '@enke.dev/lint/prettier.config.js';
 export default config;
 ```
 
-## Prepare Stylelint config (experimental)
+### Prepare Stylelint config (optional)
 
-Uses some common presets and can be used in CSS, SASS and SCSS files.\
-It will enforce a specific property order and specific custom property prefixes if configured.
+Uses common presets and can be used in CSS, SASS and SCSS files. Enforces specific property order and custom property prefixes.
 
-As this is totally opt-in, all necessary dependencies must be installed first:
+Install the required dependencies:
 
 ```bash
-npm i -D @enke.dev/lint stylelint stylelint-config-rational-order stylelint-config-standard-scss stylelint-order
+npm i -D stylelint stylelint-config-rational-order stylelint-config-standard-scss stylelint-order
 ```
 
-Create a `stylelint.config.js` file in the root of your project and add the following content:
+Create a `stylelint.config.js` file in the root of your project:
 
 ```js
 // @ts-check
-
 import { defineConfig } from '@enke.dev/lint/stylelint.config.js';
 
 export default defineConfig({ cssCustomPropertyPrefix: 'your-prefix' });
 ```
 
-For now, [no TypeScript support](https://github.com/stylelint/stylelint/issues/4940) is possible.
+> [!NOTE]
+> Stylelint [does not support TypeScript configs](https://github.com/stylelint/stylelint/issues/4940) yet.
+
+---
+
+## Approach 2: Biome (Alternative)
+
+Biome is a unified, Rust-based linter and formatter that's 10-20x faster than Node.js tools. It replaces ESLint + Prettier + Stylelint with a single tool.
+
+### Install packages
+
+```bash
+npm i -D @enke.dev/lint @biomejs/biome
+```
+
+### Prepare Biome config
+
+Create a `biome.json` file in the root of your project:
+
+```json
+{
+  "$schema": "https://biomejs.dev/schemas/latest/schema.json",
+  "extends": ["@enke.dev/lint/biome.json"]
+}
+```
+
+#### Extending the config
+
+You can override or add rules by extending the base configuration:
+
+```json
+{
+  "$schema": "https://biomejs.dev/schemas/latest/schema.json",
+  "extends": ["@enke.dev/lint/biome.json"],
+  "linter": {
+    "rules": {
+      "nursery": {
+        "noDuplicateCustomProperties": "error"
+      }
+    }
+  }
+}
+```
+
+For restricting imports (similar to ESLint example above), Biome uses a different approach:
+
+```json
+{
+  "$schema": "https://biomejs.dev/schemas/latest/schema.json",
+  "extends": ["@enke.dev/lint/biome.json"],
+  "linter": {
+    "rules": {
+      "style": {
+        "noRestrictedImports": {
+          "level": "error",
+          "options": {
+            "paths": {
+              "@/internal": "Import from @/public/* instead"
+            }
+          }
+        }
+      }
+    }
+  }
+}
+```
+
+### Limitations
+
+Biome does not support all features from ESLint/Prettier/Stylelint. See [BIOME_COMPATIBILITY.md](./BIOME_COMPATIBILITY.md) for details.
+
+**Notable limitations:**
+- No SCSS support (only standard CSS)
+- No Web Components or Lit-specific linting
+- Limited HTML linting (experimental support)
+- No plugin system (yet)
+
+**When to use Biome:**
+- Standard JavaScript/TypeScript/JSON/CSS projects
+- Performance-critical environments
+- Projects without SCSS or Web Components
+
+**When to use ESLint + Prettier + Stylelint:**
+- Projects using SCSS
+- Projects using Web Components or Lit
+- Projects requiring extensive HTML linting
+- Projects needing specific ESLint plugins
+
+---
 
 ## Development
 
-This repo self-tests the configuration by linting itself: `npm run lint`.\
-Therefore, a `test.eslint.config.ts` is used.
+This repo self-tests the configuration by linting itself: `npm run lint`
 
-And additionally, a naive test is in place to check that the linter actually finds issues: `npm run test`.\
-It uses the native Node test runner against some obviously faulty code in the `test` directory.
+A naive test suite verifies that linters detect issues: `npm run test`\
+It uses the native Node test runner against faulty code in the `test` directory.
